+++ src/payment/strategies/cybersource/index.ts
@@ -2,3 +2,6 @@ export * from './cybersource';
 
 export { default as CyberSourcePaymentStrategy } from './cybersource-payment-strategy';
 export { default as CyberSourceScriptLoader } from './cybersource-script-loader';
+export { default as CyberSourceThreeDSecurePaymentProcessor } from './cybersource-threedsecure-payment-processor';
+export { default as CyberSourcePaymentProcessor } from './cybersource-payment-processor';
+
